+++ native/java/org/jpype/html/Html.java
@@ -25,7 +25,6 @@ import java.util.HashMap;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import org.jpype.JPypeContext;
 import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
 
@@ -72,8 +71,8 @@ public class Html
 
   static
   {
-    try (InputStream is = JPypeContext.getInstance().getClass().getClassLoader()
-            .getResourceAsStream("org/jpype/html/entities.txt");
+    ClassLoader cl = ClassLoader.getSystemClassLoader();
+    try (InputStream is = cl.getResourceAsStream("org/jpype/html/entities.txt");
             InputStreamReader isr = new InputStreamReader(is);
             BufferedReader rd = new BufferedReader(isr))
     {