+++ setupext/build_ext.py
@@ -300,7 +300,7 @@ class BuildExtCommand(build_ext):
         distutils.log.info(
             "Jar cache is missing, using --enable-build-jar to recreate it.")
 
-        target_version = "1.8"
+        target_version = "9"
         # build the jar
         try:
             dirname = os.path.dirname(self.get_ext_fullpath("JAVA"))
@@ -311,10 +311,16 @@ class BuildExtCommand(build_ext):
             cmd1 = shlex.split('%s -cp "%s" -d "%s" -g:none -source %s -target %s -encoding UTF-8' %
                                (javac, classpath, build_dir, target_version, target_version))
             cmd1.extend(ext.sources)
-
-            os.makedirs("build/classes", exist_ok=True)
             self.announce("  %s" % " ".join(cmd1), level=distutils.log.INFO)
             subprocess.check_call(cmd1)
+
+            cmd1 = shlex.split('%s -cp "%s" -d "%s" -g:none -source %s -target %s -encoding UTF-8' %
+                               (javac, build_dir, os.path.join(build_dir,"META-INF","versions","0"), target_version, target_version))
+            cmd1.extend(glob.glob(os.path.join("native","java0","**","*.java"), recursive=True))
+            os.makedirs(os.path.join(build_dir,"META-INF","versions","0"), exist_ok=True)
+            self.announce("  %s" % " ".join(cmd1), level=distutils.log.INFO)
+            subprocess.check_call(cmd1)
+
             manifest = None
             try:
                 for file in glob.iglob("native/java/**/*.*", recursive=True):